# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "车险报价对比助手 API",
    "description": "MVP 后端接口；FastAPI 自动生成的 OpenAPI 是唯一请求/响应契约。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "description": "项目列表（含报价数与最低有效净支出聚合；无报价时为稳定空列表）。",
        "operationId": "list_projects_api_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_ProjectListItem__"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "创建续保对比项目（新建项目报价分组为稳定空列表）。",
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProjectDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "description": "项目详情（含按“公司+保险员”分组的报价卡数据）。",
        "operationId": "get_project_api_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProjectDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project",
        "description": "编辑项目基础信息。",
        "operationId": "update_project_api_projects__project_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProjectDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "description": "删除项目（不可恢复）：级联删除报价与文件记录，并预约磁盘清理。",
        "operationId": "delete_project_api_projects__project_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_NoneType_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "type": "null",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiResponse[NoneType]"
      },
      "ApiResponse_ProjectDetail_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectDetail"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiResponse[ProjectDetail]"
      },
...
      "ApiResponse_list_ProjectListItem__": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProjectListItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiResponse[list[ProjectListItem]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NetPaymentStatus": {
        "type": "string",
        "enum": [
          "OK",
          "MISSING_TOTAL",
          "INVALID_DISCOUNT"
        ],
        "title": "NetPaymentStatus",
        "description": "净支出状态：缺失总价 / 优惠超额都不得自行算 0。"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "项目名称，如“2026 车辆续保”"
          },
          "vehicleName": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Vehiclename",
            "description": "车辆名称，如 Model Y"
          },
          "renewalYear": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "Renewalyear",
            "description": "续保年份"
          },
          "expireDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiredate",
            "description": "保险到期时间（可选）"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "备注（可选，入库前脱敏）"
          }
        },
        "type": "object",
        "required": [
          "name",
          "vehicleName",
          "renewalYear"
        ],
        "title": "ProjectCreate",
        "description": "创建项目：字段严格限定为项目名、车辆名称、续保年份、可选到期日与备注。"
      },
      "ProjectDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "vehicleName": {
            "type": "string",
            "title": "Vehiclename"
          },
          "renewalYear": {
            "type": "integer",
            "title": "Renewalyear"
          },
          "expireDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiredate"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "vehicleModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehiclemodel"
          },
          "vehicleSeats": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicleseats"
          },
          "firstRegDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstregdate"
          },
          "isNev": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isnev"
          },
          "modelConsentAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelconsentat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "quoteGroups": {
            "items": {
              "$ref": "#/components/schemas/QuoteGroup"
            },
            "type": "array",
            "title": "Quotegroups"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "vehicleName",
          "renewalYear",
          "createdAt",
          "updatedAt",
          "quoteGroups"
        ],
        "title": "ProjectDetail",
        "description": "项目详情：在基础字段之上附带分组报价卡数据（必填集合）。"
      },
      "ProjectListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "vehicleName": {
            "type": "string",
            "title": "Vehiclename"
          },
          "renewalYear": {
            "type": "integer",
            "title": "Renewalyear"
          },
          "expireDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiredate"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "vehicleModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehiclemodel"
          },
          "vehicleSeats": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicleseats"
          },
          "firstRegDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstregdate"
          },
          "isNev": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isnev"
          },
          "modelConsentAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelconsentat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "quoteCount": {
            "type": "integer",
            "title": "Quotecount"
          },
          "minNetPayment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minnetpayment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "vehicleName",
          "renewalYear",
          "createdAt",
          "updatedAt",
          "quoteCount"
        ],
        "title": "ProjectListItem",
        "description": "项目列表项：附带报价数与最低有效净支出聚合。\n\n“有效净支出”只统计已确认（CONFIRMED/MERGE_REVIEW）且 netPayment 非空的报价；\n无有效报价时 minNetPayment 为 null（稳定空状态，前端不得当作 0 处理）。"
      },
      "ProjectUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "vehicleName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehiclename"
          },
          "renewalYear": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 2000.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Renewalyear"
          },
          "expireDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiredate"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "ProjectUpdate",
        "description": "编辑项目：全部字段可选，仅更新显式提供的字段（PATCH 语义）。"
      },
      "QuoteCardSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "insurerCode": {
            "type": "string",
            "title": "Insurercode"
          },
          "insurerName": {
            "type": "string",
            "title": "Insurername"
          },
          "agentName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agentname"
          },
          "planLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planlabel"
          },
          "source": {
            "$ref": "#/components/schemas/QuoteSource"
          },
          "status": {
            "$ref": "#/components/schemas/QuoteStatus"
          },
          "netPayment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Netpayment"
          },
          "netPaymentStatus": {
            "$ref": "#/components/schemas/NetPaymentStatus"
          },
          "officialTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Officialtotal"
          },
          "computedTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Computedtotal"
          },
          "totalCheckStatus": {
            "$ref": "#/components/schemas/TotalCheckStatus"
          },
          "thirdPartyAmount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thirdpartyamount"
          },
          "tpNonMedicalAmount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tpnonmedicalamount"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "insurerCode",
          "insurerName",
          "source",
          "status",
          "netPaymentStatus",
          "totalCheckStatus",
          "createdAt"
        ],
        "title": "QuoteCardSummary",
        "description": "项目详情页报价卡片摘要（分组内的单张卡）。\n\n净支出为 null 时按 netPaymentStatus 区分“总价缺失/优惠超额”，\n前端不得把 null 当 0；totalCheckStatus=MISMATCH 必须展示异常提示。"
      },
...
      "QuoteGroup": {
        "properties": {
          "insurerCode": {
            "type": "string",
            "title": "Insurercode"
          },
          "insurerName": {
            "type": "string",
            "title": "Insurername"
          },
          "agentName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agentname"
          },
          "sameSourceHint": {
            "type": "boolean",
            "title": "Samesourcehint"
          },
          "quotes": {
            "items": {
              "$ref": "#/components/schemas/QuoteCardSummary"
            },
            "type": "array",
            "title": "Quotes"
          }
        },
        "type": "object",
        "required": [
          "insurerCode",
          "insurerName",
          "sameSourceHint",
          "quotes"
        ],
        "title": "QuoteGroup",
        "description": "按“保险公司 + 保险员”分组的报价卡组（SPEC §8 / 决策 #9）。\n\n组内多于一份报价时 sameSourceHint=true：仅提示“同来源报价”，\n不创建版本链（MVP 无 QuoteVersion）。quotes 无默认（构造时必填）。"
      },
...
      "QuoteSource": {
        "type": "string",
        "enum": [
          "UPLOADED",
          "MANUAL"
        ],
        "title": "QuoteSource",
        "description": "报价来源：上传解析 / 纯手动录入。"
      },
      "QuoteStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "PARSING",
          "PENDING_CONFIRM",
          "CONFIRMED",
          "PARSE_FAILED",
          "MERGE_REVIEW"
        ],
        "title": "QuoteStatus",
        "description": "报价状态机（SPEC §2.10），MVP 冻结不允许改动取值。"
      },
...
      "TotalCheckStatus": {
        "type": "string",
        "enum": [
          "NOT_CHECKABLE",
          "PASSED",
          "MISMATCH"
        ],
        "title": "TotalCheckStatus",
        "description": "总额校验三态，避免把无法校验误写为“通过”。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}